funi/requirements.txt
torch
transformers
accelerate
python-dotenv
discord.py
selenium
googlesearch-python
datasets
gradio